requirements.txt
python-telegram-bot==10.0.1
paramiko==2.4.1
pytz
